Fix URLs in modern Blogger backups
Posts use https:// schema in their "alternate" links, but comments refer
to posts via http://. This breaks the program's understanding of the
relationships.

To work around this, I "upgrade" all links to HTTPS. This is a kludge;
a better way is probably to look at tbh:in-reply-to tag on comments, but
the `feed` library doesn't pass that info through.

Fixes #32.